refactor: frame source, r3d package, browser path, packaging
Split optional R3D into src/core/r3d/, unify video/R3D convert and player
decode via frame_source, extract browser path paste helpers, and harden
release packaging (no strip under r3d/, shared safe_print/junk filters).

Changelog: 0.9.6 notes.
